organize common functions
