+++ common/io.go
@@ -1,0 +1,10 @@
+package common
+
+import "os"
+
+func FileExists(path string) bool {
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		return false
+	}
+	return true
+}
